+++ src/sfptpd_metrics.c
@@ -1515,6 +1515,17 @@ int sfptpd_metrics_listener_open(struct sfptpd_config *app_config)
 	        goto fail;
 	}
 
+	/* Set ownership of socket. Defer error to any consequent failure. */
+	if (fchown(metrics.listen_fd, general_config->uid, general_config->gid))
+		TRACE_L4(PREFIX "could not set socket ownership, %s\n",
+			 strerror(errno));
+
+	/* Set access mode. Be louder because this is explicit config. */
+	if (general_config->metrics_socket_mode != (mode_t) -1 &&
+	    fchmod(metrics.listen_fd, general_config->metrics_socket_mode) == -1)
+		WARNING(PREFIX "could not set socket mode, %s\n",
+			strerror(errno));
+
 	/* Bind to the path in the filesystem. */
 	rc = bind(metrics.listen_fd, (const struct sockaddr *) &addr, sizeof addr);
 	if (rc == -1) {
@@ -1522,11 +1533,6 @@ int sfptpd_metrics_listener_open(struct sfptpd_config *app_config)
 		      metrics_path);
 	        goto fail;
 	}
-
-	/* Set ownership of socket. Defer error to any consequent failure. */
-	if (chown(metrics_path, general_config->uid, general_config->gid))
-		WARNING(PREFIX "could not set socket ownership, %s\n",
-			 strerror(errno));
 
 	flags = fcntl(metrics.listen_fd, F_GETFL);
 	rc = fcntl(metrics.listen_fd, F_SETFL, flags | O_NONBLOCK);